requirements/base.txt
Django==5.2.2
djangorestframework==3.17.1
drf-nested-routers==0.95.0
django-filter==25.2
python-dotenv==1.1.0
